requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
